+++ server/src/constants/cadFiles.js
@@ -1,0 +1,10 @@
+export const MODEL_FILE_EXTENSIONS = Object.freeze([
+  '.stp',
+  '.step',
+  '.stl',
+  '.iges',
+  '.igs',
+  '.wrl',
+  '.3ds',
+  '.x_t',
+]);